feat: add authenticated URL delete endpoint
