+++ collision.go
@@ -1,0 +1,28 @@
+package main
+
+import (
+	"context"
+	"database/sql"
+
+	"url-short/internal/database"
+	"url-short/internal/shortener"
+)
+
+func hashCollisionDetection(DB *database.Queries, url string, count int, requestContext context.Context) (string, error) {
+	hashURL := shortener.Hash(url, count)
+	shortURLHash := shortener.Shorten(hashURL)
+
+	_, err := DB.SelectURL(requestContext, shortURLHash)
+
+	if err == sql.ErrNoRows {
+		return shortURLHash, nil
+	}
+
+	if err != nil && err != sql.ErrNoRows {
+		return "", err
+	}
+
+	count++
+
+	return hashCollisionDetection(DB, url, count, requestContext)
+}
