+++ main.go
@@ -41,6 +41,8 @@ func main() {
 	// url management endpoints
 	mux.HandleFunc("POST /api/v1/data/shorten", apiCfg.authenticationMiddlewear(apiCfg.postLongURL))
 	mux.HandleFunc("GET /api/v1/{shortUrl}", apiCfg.getShortURL)
+	mux.HandleFunc("DELETE /api/v1/{shortUrl}", apiCfg.authenticationMiddlewear(apiCfg.deleteShortURL))
+
 
 	// user management endpoints
 	mux.HandleFunc("POST /api/v1/users", apiCfg.postAPIUsers)
